# docs/api-reference/swagger.yaml
openapi: "3.0.3"
info:
  description: Sandbox API
  version: 1.0.0
  title: Sandbox API
  license:
    name: MIT
servers:
  - url: /api/v1
...
paths:
...
  /dns-account-configurations:
    post:
      summary: Create a new DNSAccountConfiguration
      operationId: createDNSAccountConfiguration
      tags:
        - admin
      requestBody:
        description: JSON object to specify the DNSAccountConfiguration.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DNSAccountConfiguration"
            example:
              name: devaccount
              aws_access_key_id: redacted
              aws_secret_access_key: redacted
              zone: devzone
              hosted_zone_id: Z0
              annotations:
                purpose: dev
      responses:
        '201':
          description: DNSAccountConfiguration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                message: DNS account configuration created
        '400':
          description: Wrong request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: Bad request
                http_code: 400
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: unauthorized
                http_code: 401
        default:
          description: createDNSAccountConfiguration unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Get all DNSAccountConfigurations
      operationId: getDNSAccountConfigurations
      tags:
        - admin
      responses:
        '200':
          description: The list of DNSAccountConfigurations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DNSAccountConfigurations"
        default:
          description: getDNSAccountConfigurations unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Annotations:
      description: Key / Value map to provide optional information.
      type: object
      additionalProperties:
        type: string
        minLength: 1
      example:
        guid: abcde
        env_type: ocp4-cluster
        hcp: "yes"
...
    Error:
      description: Error response with HTTP code and message
      type: object
      required:
        - http_code
        - message
      properties:
        http_code:
          type: integer
          format: integer
          example: 502
        message:
          type: string
          example: Unexpected Error
    Message:
      description: Simple message response
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Account successfully marked for cleanup
...
    DNSAccountConfiguration:
      description: >
        DNSAccountConfiguration is a record of a DNS Account.
        provider_type determines the backend: 'route53' (AWS) or 'ddns' (BIND TSIG key API).
      type: object
      required:
        - name
        - zone
        - annotations
      additionalProperties: false
      properties:
        name:
          type: string
          description: The name of the account, must be alphanumeric or dash
          example: devaccount
        provider_type:
          type: string
          description: Backend provider type
          enum: [route53, ddns]
          default: route53
          example: route53
        aws_access_key_id:
          type: string
          description: The access key id to access the AWS Account (required for route53)
          example: AKI..
        aws_secret_access_key:
          type: string
          description: The secret access key to access the AWS Account (required for route53)
          example: ey..
        valid:
          type: boolean
          description: true if the cluster is valid
          example: true
          default: true
        zone:
          type: string
          example: dev.domain
        hosted_zone_id:
          type: string
          description: AWS hosted zone id (required for route53)
          example: Z0
        endpoint:
          type: string
          description: URL of the BIND TSIG key API (required for ddns)
          example: https://ddns.example.com:8080
        token:
          type: string
          description: Bearer auth token for the BIND TSIG key API (required for ddns)
          example: my-secret-token
        annotations:
          $ref: "#/components/schemas/Annotations"
        additional_vars:
          type: object
          description: Additional variables for the cluster
          example:
            deployer:
              openshift_cnv_nfs_path: /IBMfoobar/data01
              openshift_cnv_nfs_server: fsf-region.domain.com
      example:
        name: devaccount
        provider_type: route53
        aws_access_key_id: REDACTED
        aws_secret_access_key: REDACTED
        zone: dev.domain.com
        hosted_zone_id: Z0redacted
        annotations:
          purpose: dev
    DNSAccountConfigurations:
      description: The list of DNSAccountConfigurations
      type: array
      items:
        $ref: '#/components/schemas/DNSAccountConfiguration'
...